Gestion de l'autorisation dans le controleur (voir #37#issuecomment-5534517) + ajout des methodes de PromoController dans PromosController
